Enhance colocation by restructuring folders
- Set the 'core' directory as the container for all domain-related code
- Establish the 'index.ts' file in the 'user' directory as the entry point for the user domain, exporting domain types and functions
- Introduce the 'usecases' subdirectory within the 'user' directory to holds use cases such as createUser and getUser
- Establish the 'index.ts' file in the 'usecases' directory as the entry point for user use cases, exporting use case interfaces
- Introduce the 'adapters' subdirectory within the 'user' directory to holds adapters such as userStorage
- Merge the 'useUserStorage.ts' file into the 'store.ts' file, eliminating unnecessary 'hook' directory
- Apply similar directory structures for other domains such as 'achievement', 'challenge', and 'enrollment' to group their respective files
