--- app/hooks/useUserStorage.ts
@@ -1,8 +1,0 @@
-import { userStorage } from '@/adapters/userStorage'
-
-import { useUserStore } from '@/app/services/store'
-
-export const useUserStorage = () => {
-  const userStore = useUserStore()
-  return userStorage(userStore)
-}
