+++ app/services/store.ts
@@ -1,9 +1,15 @@
-import { type User } from '@/domain/user'
-import { type UserStorage } from '@/usecases/ports'
+import { type User } from '@/core/user'
+import { type UserStorage } from '@/core/user/usecases'
+import { userStorage } from '@/core/user/adapters/userStorage'
 
 import { create } from 'zustand'
 
-export const useUserStore = create<UserStorage>((set) => ({
+const useUserStore = create<UserStorage>((set) => ({
   user: undefined,
   createUser: (user: User) => set({ user }),
 }))
+
+export const useUserStorage = () => {
+  const userStore = useUserStore()
+  return userStorage(userStore)
+}
