+++ core/achievement/background.d.ts
@@ -1,0 +1,11 @@
+type Background = SolidBackground | EmojiBackground;
+
+type SolidBackground = {
+  type: "solid";
+  color: string;
+};
+
+type EmojiBackground = {
+  type: "emoji";
+  char: string;
+};
