+++ core/user/usecases/getUser.ts
@@ -1,0 +1,11 @@
+import { type UserStorage } from '@/core/user/usecases'
+
+type Dependencies = {
+  userStorage: UserStorage
+}
+
+export const getUser = (dependencies: Dependencies) => {
+  const { userStorage } = dependencies
+
+  return userStorage.user
+}
