+++ core/user/usecases/index.ts
@@ -1,0 +1,6 @@
+import { type User } from '@/core/user'
+
+export interface UserStorage {
+  user?: User
+  createUser(user: User): void
+}
