--- domain/background.d.ts
@@ -1,11 +1,0 @@
-type Background = SolidBackground | EmojiBackground;
-
-type SolidBackground = {
-  type: "solid";
-  color: string;
-};
-
-type EmojiBackground = {
-  type: "emoji";
-  char: string;
-};
