--- usecases/getUser.ts
@@ -1,11 +1,0 @@
-import { type UserStorage } from '@/usecases/ports'
-
-type Dependencies = {
-  userStorage: UserStorage
-}
-
-export const getUser = (dependencies: Dependencies) => {
-  const { userStorage } = dependencies
-
-  return userStorage.user
-}
