--- usecases/ports.ts
@@ -1,6 +1,0 @@
-import { type User } from '@/domain/user'
-
-export interface UserStorage {
-  user?: User
-  createUser(user: User): void
-}
